[src] widget: use `u16` in size hints

+++ src/widget/mod.rs
@@ -9,9 +9,9 @@ pub use text::Text;
 use crate::{Area, Message};
 
 pub trait Widget {
-	fn get_width_hint(&self) -> usize;
+	fn get_width_hint(&self) -> u16;
 
-	fn get_height_hint(&self) -> usize;
+	fn get_height_hint(&self) -> u16;
 
 	fn render(&self, area: Area) -> String;
